refactor: rename src/task/action-types.ts to types.ts
