+++ src/redux-states/task/index.ts
@@ -1,6 +1,6 @@
 import * as taskActions from './actions';
 import * as taskConstants from './action-constants';
-import * as taskTypes from './action-types';
+import * as taskTypes from './types';
 import taskReducer from './reducer';
 
 export { taskActions, taskConstants, taskReducer, taskTypes };
